Add files via upload
Libraries Used:

face_recognition: This library provides the core functionality for facial recognition.
cv2 (OpenCV): OpenCV is used for camera access and image processing.
numpy: NumPy is used for numerical operations.
os: It is used for file and directory operations.
pyttsx3: This library is used for text-to-speech functionality.
Voice Alert:

The code initializes the text-to-speech engine using the pyttsx3 library and sets the voice property.
The speak(audio) function is defined to provide a voice alert with the specified audio message.
Known Faces:

A list named list is created, which contains names of known individuals. If any of these individuals are detected, the code will trigger a voice alert.
Training Data:

The code assumes that the images of known individuals are stored in the "./train/" directory.
It loads these images, encodes their faces, and stores them in known_name_encodings.
Corresponding names are stored in the known_names list.
Real-Time Face Recognition:

The code continuously captures frames from the computer's webcam.
It detects face locations and encodes them in the captured frames.
It compares the encoded faces with the known face encodings.
If a match is found, it determines the name of the detected individual and checks if that name is in the list of known individuals.
If a known individual is detected and listed in the list, a voice alert is triggered, indicating a "criminal detected."
If the detected face doesn't match any known individual, it is labeled as "unknown."
Visual Feedback:

The code draws rectangles around detected faces and displays the name or "unknown" label on the video frame.
It uses OpenCV to show the video feed in a window.